login pages and changes

+++ src/main/webapp/WEB-INF/view/list-users.jsp
@@ -1,34 +1,21 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 
 <!DOCTYPE html>
-
 <html>
-
 <head>
-	<title>List Users</title>
-	
+	<title>List of Employees</title>
 	<!-- reference our style sheet -->
-
-	<link type="text/css"
-		  rel="stylesheet"
-		  href="${pageContext.request.contextPath}/resources/css/style.css" />
-
+	<link type="text/css" rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/style.css" />
 </head>
-
 <body>
-
 	<div id="wrapper">
 		<div id="header">
 			<h2>Welcome Here</h2>
 		</div>
 	</div>
-	
 	<div id="container">
-	
 		<div id="content">
-		
 			<!-- put new button: Add Customer -->
-		
 			<input type="button" value="Add User" onclick="window.location.href='showFormForAdd'; return false;" class="add-button"/>
 		
 			<!--  add our html table here -->
@@ -65,23 +52,15 @@
 						<td> ${tempUser.phoneNumber} </td>					
 						<td>
 							<!-- display the update link -->
-							<a href="${updateLink2}">Update</a>
-							|
-							<a href="${deleteLink2}"
-							   onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>
+							<a href="${updateLink2}">Update</a> |
+							<a href="${deleteLink2}" onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>
 						</td>
-						
 					</tr>
-				
-				</c:forEach>
-						
-			</table>
-				
+				</c:forEach>	
+			</table>	
 		</div>
-	
 	</div>
-	
-
+	<br><br>
+	<a href="${pageContext.request.contextPath}/">--Back to home</a>
 </body>
-
 </html>
